Made ResetVisibility and ParseRepo generic. Validate new from repo has admin rights
